General warnings removal.

+++ src/tests/s1c3_sbxor_decode_test.c
@@ -4,7 +4,7 @@
 void print_chars(const unsigned char *in, const unsigned int len)
 {
     int i;
-    for (i=0; i<len; i++)
+    for (i=0; (unsigned int)i<len; i++)
         printf("%c", in[i]);    
 }
