+++ src/util/printbuffer.c
@@ -1,15 +1,16 @@
+#include <stdio.h>
 #include "printbuffer.h"
 
 void print_chars(const unsigned char *in, const unsigned int len)
 {
     int i;
-    for (i=0; i<len; i++)
+    for (i=0; (unsigned int)i<len; i++)
         printf("%c", in[i]);    
 }
 
 void print_hex(const unsigned char *in, const unsigned int len)
 {
     int i;
-    for (i=0; i<len; i++)
+    for (i=0; (unsigned int)i<len; i++)
         printf("%02.2X", in[i]);
 }
